+++ Nbuild/Command.cs
@@ -1,8 +1,10 @@
-﻿using NbuildTasks;
+﻿using CommandLine;
+using NbuildTasks;
 using Ntools;
 using OutputColorizer;
 using System.Diagnostics;
 using System.Runtime.InteropServices;
+using System.Text.Json;
 
 namespace Nbuild
 {
@@ -28,6 +30,7 @@ namespace Nbuild
             DownloadsDirectory = !TestMode || Ntools.CurrentProcess.IsElevated() ? "C:\\NToolsDownloads" : $"{Environment.GetEnvironmentVariable("Temp")}\\nb";
 
             if (!Directory.Exists(DownloadsDirectory)) Directory.CreateDirectory(DownloadsDirectory);
+
         }
 
         private static bool IsTestMode()
@@ -92,85 +95,68 @@ namespace Nbuild
             return true;
         }
 
-        public static ResultHelper Install(string json, bool verbose = false)
+        public static ResultHelper Install(string? json, bool verbose = false)
         {
             Verbose = verbose;
-            // check if caller is admin or in test mode
+            ResultHelper result = ResultHelper.New();
             if (!CanRunCommand()) return ResultHelper.Fail(-1, $"You must run this command as an administrator");
 
-            try
+
+            var appDataList = GetApps(json);
+            if (appDataList == null) return ResultHelper.Fail(-1, $"Invalid json input");
+
+            if (Verbose) Colorizer.WriteLine($"[{ConsoleColor.Yellow}!{appDataList.Count()} apps to install.]");
+
+            foreach (var appData in appDataList)
             {
-                var appDataList = NbuildApp.GetApps(json);
-                if (appDataList == null) return ResultHelper.Fail(-1, $"Invalid json input");
-
-                if (Verbose) Colorizer.WriteLine($"[{ConsoleColor.Yellow}!{appDataList.Count()} apps to install.]");
-
-                foreach (var appData in appDataList)
+                result = Install(appData);
+                if (!result.IsSuccess())
                 {
-                    var result = Install(appData);
-                    if (!result.IsSuccess())
-                    {
-                        return result;
-                    }
+                    break;
                 }
-
-                Console.WriteLine();
-                // all apps installed successfully
-                return ResultHelper.Success();
             }
-            catch (Exception ex)
-            {
-                Console.WriteLine();
-                return ResultHelper.Fail(-1, $"Invalid json input: {ex.Message}");
 
-            }
+            return result;
         }
 
-        public static ResultHelper List(string json, bool verbose = false)
+        public static ResultHelper List(string? json, bool verbose = false)
         {
             Verbose = verbose;
-            try
+
+            var appDataList = GetApps(json);
+
+            if (appDataList == null) return ResultHelper.Fail(-1, $"Invalid json input");
+
+            Colorizer.WriteLine($"[{ConsoleColor.Yellow}!{appDataList.Count()} apps to list:]");
+
+            // print header
+            Colorizer.WriteLine($"[{ConsoleColor.Yellow}!|--------------------|----------------|-------------------|]");
+            Colorizer.WriteLine($"[{ConsoleColor.Yellow}!| App name           | Target version | Installed version |]");
+            Colorizer.WriteLine($"[{ConsoleColor.Yellow}!|--------------------|----------------|-------------------|]");
+            foreach (var appData in appDataList)
             {
-                var appDataList = NbuildApp.GetApps(json);
-                if (appDataList == null) return ResultHelper.Fail(-1, $"Invalid json input");
-
-                Colorizer.WriteLine($"[{ConsoleColor.Yellow}!{appDataList.Count()} apps to list:]");
-
-                // print header
-                Colorizer.WriteLine($"[{ConsoleColor.Yellow}!|--------------------|----------------|-------------------|]");
-                Colorizer.WriteLine($"[{ConsoleColor.Yellow}!| App name           | Target version | Installed version |]");
-                Colorizer.WriteLine($"[{ConsoleColor.Yellow}!|--------------------|----------------|-------------------|]");
-                foreach (var appData in appDataList)
+                // display app and installed version
+                // InstalledAppFileVersionGreterOrEqual is true, print green, else print red
+                if (InstalledAppFileVersionGreaterOrEqual(appData))
                 {
-                    // display app and installed version
-                    // InstalledAppFileVersionGreterOrEqual is true, print green, else print red
-                    if (InstalledAppFileVersionGreaterOrEqual(appData))
-                    {
-                        Colorizer.WriteLine($"[{ConsoleColor.Green}!| {appData.Name,-18} | {appData.Version,-14} | {GetNbuildAppFileVersion(appData),-18}|]");
-                    }
-                    else
-                    {
-                        Colorizer.WriteLine($"[{ConsoleColor.Red}!| {appData.Name,-18} | {appData.Version,-14} | {GetNbuildAppFileVersion(appData),-18}|]");
-                    }
+                    Colorizer.WriteLine($"[{ConsoleColor.Green}!| {appData.Name,-18} | {appData.Version,-14} | {GetNbuildAppFileVersion(appData),-18}|]");
                 }
-
-                Console.WriteLine();
-            }
-            catch (Exception ex)
-            {
-                return ResultHelper.Fail(-1, ex.Message);
+                else
+                {
+                    Colorizer.WriteLine($"[{ConsoleColor.Red}!| {appData.Name,-18} | {appData.Version,-14} | {GetNbuildAppFileVersion(appData),-18}|]");
+                }
             }
 
+            Console.WriteLine();
             return ResultHelper.Success();
         }
 
-        public static ResultHelper Download(string json, bool verbose = false)
+        public static ResultHelper Download(string? json, bool verbose = false)
         {
             Verbose = verbose;
-            // check if admin
-            if (!CanRunCommand()) return ResultHelper.Fail(-1, $"You must run this command as an administrator");
+            
+            var appDataList = GetApps(json);
 
-            var appDataList = NbuildApp.GetApps(json);
             if (appDataList == null) return ResultHelper.Fail(-1, $"Invalid json input");
 
             Colorizer.WriteLine($"[{ConsoleColor.Yellow}!{appDataList.ToList().Count} apps to download to {DownloadsDirectory}]");
@@ -348,6 +334,107 @@ namespace Nbuild
 
             //if (Verbose) Colorizer.WriteLine($"[{ConsoleColor.Yellow}!{nbuildApp.Name} current version: {currentVersion} >=  {nbuildApp.Version}: {currentVersionParsed >= versionParsed}]");
             return currentVersionParsed >= versionParsed;
+        }
+
+        public static IEnumerable<NbuildApp> GetApps(string? json)
+        {
+            if (string.IsNullOrEmpty(json))
+            {
+                throw new ArgumentNullException(nameof(json));
+            }
+
+            // check if json is a file path
+            if (File.Exists(json))
+            {
+                json = File.ReadAllText(json);
+
+                if (string.IsNullOrEmpty(json))
+                {
+                    throw new ArgumentNullException(nameof(json));
+                }
+            }
+
+            var listAppData = JsonSerializer.Deserialize<NbuildApps>(json) ?? throw new ParserException("Failed to parse json to list of objects", null);
+
+            // make sure version matches 1.2.0
+            if (listAppData.Version != "1.2.0")
+            {
+                throw new ParserException($"Version {listAppData.Version} is not supported. Please use version 1.2.0", null);
+            }
+
+            foreach (var appData in listAppData.NbuildAppList)
+            {
+                yield return Validate(appData);
+            }
+
+        }
+
+        private static NbuildApp Validate(NbuildApp appData)
+        {
+            if (string.IsNullOrEmpty(appData.Name))
+            {
+                throw new ParserException("Name is missing or empty", null);
+            }
+            if (string.IsNullOrEmpty(appData.WebDownloadFile))
+            {
+                throw new ParserException("WebDownloadFile is missing or empty", null);
+            }
+            if (string.IsNullOrEmpty(appData.DownloadedFile))
+            {
+                throw new ParserException("DownloadedFile is missing or empty", null);
+            }
+            if (string.IsNullOrEmpty(appData.InstallCommand))
+            {
+                throw new ParserException("InstallCommand is missing or empty", null);
+            }
+            if (string.IsNullOrEmpty(appData.InstallArgs))
+            {
+                throw new ParserException("InstallArgs is missing or empty", null);
+            }
+            if (string.IsNullOrEmpty(appData.Version))
+            {
+                throw new ParserException("Version is missing or empty", null);
+            }
+            if (string.IsNullOrEmpty(appData.InstallPath))
+            {
+                throw new ParserException("InstallPath is missing or empty", null);
+            }
+            if (string.IsNullOrEmpty(appData.AppFileName))
+            {
+                throw new ParserException("AppFileName is missing or empty", null);
+            }
+
+
+            // Replace $(Version) with the actual version in appData.Version
+            // Replace $(InstallPath) with the actual version in appData.InstallPath
+            appData.WebDownloadFile = appData.WebDownloadFile.Replace("$(Version)", appData.Version);
+
+            appData.DownloadedFile = appData.DownloadedFile.Replace("$(Version)", appData.Version);
+            appData.DownloadedFile = appData.DownloadedFile.Replace("$(AppFileName)", appData.AppFileName);
+
+            appData.InstallCommand = appData.InstallCommand.Replace("$(Version)", appData.Version);
+            appData.InstallCommand = appData.InstallCommand.Replace("$(DownloadedFile)", appData.DownloadedFile);
+
+            string programFiles = Environment.GetFolderPath(Environment.SpecialFolder.ProgramFiles);
+            string programFilesX86 = Environment.GetFolderPath(Environment.SpecialFolder.ProgramFilesX86);
+
+            appData.InstallArgs = appData.InstallArgs.Replace("$(Version)", appData.Version);
+            appData.InstallArgs = appData.InstallArgs.Replace("$(InstallPath)", appData.InstallPath);
+            appData.InstallArgs = appData.InstallArgs.Replace("$(AppFileName)", appData.AppFileName);
+
+            appData.InstallArgs = appData.InstallArgs.Replace("$(DownloadedFile)", appData.DownloadedFile);
+
+            appData.InstallArgs = appData.InstallArgs.Replace("$(ProgramFiles)", programFiles);
+            appData.InstallArgs = appData.InstallArgs.Replace("$(ProgramFilesX86)", programFilesX86);
+
+            appData.InstallPath = appData.InstallPath.Replace("$(Version)", appData.Version);
+            appData.InstallPath = appData.InstallPath.Replace("$(ProgramFiles)", programFiles);
+            appData.InstallPath = appData.InstallPath.Replace("$(ProgramFilesX86)", programFilesX86);
+
+
+            return !Path.IsPathRooted(appData.InstallPath)
+                ? throw new ParserException($"App: {appData.Name}, InstallPath {appData.InstallPath} must be rooted. i.e. C:\\Program Files\\Nbuild", null)
+                : appData;
         }
     }
 }
