--- Nbuild/NbuildApps.cs
@@ -1,7 +1,0 @@
-using Nbuild;
-
-public class NbuildApps
-{
-    public required string Version { get; set; }
-    public required List<NbuildApp> NbuildAppList { get; set; }
-}
